Calculate the dynamic type of Optional's payload whenever possible
When dumping the value of an Optional, we may not calculate the dynamic
type of the payload. This patch makes it so that we always calculate
this dynamic type. The user-visible change occurs for code like this:

  let d: NSObject? = MyClass() as NSObject

Before this patch, lldb would dump the value of 'd' as:

  d: NSObject? = 0x0000000100200060

After this patch, it would instead print:

  d: NSObject? = 0x0000000100200060 {
    ObjectiveC.NSObject = {}
  }

rdar://39736324