Avoid adding duplicate configuration macros to clang importer options.
When concatenating the clang options from all dylibs we don't unique
configuration macros which can lead to more unnecessary module
configurations to be built.

This tests that configuration macros get uniqued when building the
scratch ast context. Note that "-D MACRO" options with a space are
currently only combined to "-DMACRO" when they appear outside of the
main binary. Since this commit is a performance optimization this
seems to be the best trade-off between complexity of implementation
and performance.

<rdar://problem/40220286>