Improve coverage of the apple-tables test
After closer examination, it turns out we mis-classify one of the
methods only if two of the structs have the same name. Since this was
meant to be a basic test, I rename one of the structs in the test so
that we have at least some coverage for the apple tables lookup.
Instead, I create an XFAILed test which specifically targets the
same-name case (and file a bug to track it).

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@332833 91177308-0d34-0410-b5e6-96231b3b80d8