Make test sources compatible with android+libcxx+modules
In a modules build, android is very picky about which symbols are
visible after including libc++ headers (e.g. <cstdio> defines only
std::printf and not ::printf).

This consolidates the tests where this was an issue to always include
the <c???> version of the headers and prefixes the symbols with std:: as
necessary.

Apart from that, there is no functional change in the tests.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@335149 91177308-0d34-0410-b5e6-96231b3b80d8