[CMake] Move swift compiler logic
Instead of duplicating the logic that passes the swift compiler and
library path, move it into test/CMakeLists.txt and pass it as a regular
argument to lldb-dotest.py.

This also gets rids of the remaining generator expressions which are
incompatible with the way we currently generate the lldb-dotest wrapper
and lit runner for the test suite.