Relax TestCPP11EnumTypes.py
This data formatter test broke because of a revert made inside of clang:
debug info support for certain kinds of enums was rolled back, and the
resulting loss of information prevents lldb's data formatter from
printing the expected result.

The clang functionality will eventually be un-reverted. When that
happens we can make this test strict again.

rdar://42556903