Fix deadlock in gdb-client tests
Using a listen queue of length 0 caused a deadlock on my machine in the
gdb-client tests while attempting to establish the loopback socket
connection.

I am not sure if this is down to a different python or kernel version,
but in either case, having queue of length zero sounds like a bad idea,
so I'm bumping that to one (which also fixes the deadlock).

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@341096 91177308-0d34-0410-b5e6-96231b3b80d8