Use a relaxed substring check for function names in a test
The TestTailCallFrameSBAPI.py test checks that function names in a
backtrace are equal to an expected value.

Use a relaxed substring check because function dislpay names are
platform-dependent. E.g we see "void sink(void)" on Windows, but "sink()" on
Darwin. This seems like a bug -- just work around it for now.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@344634 91177308-0d34-0410-b5e6-96231b3b80d8