[Swift] Fix several bugs in existential dynamic type resolution.
lldb was performing their own dynamic type resolution and payload
projection. At lesat the following cases were broken (and fixed
with my patch), in a generic context:

-> Error as AnyObject (class-backed existentials)
-> enums conforming to the Error protocol (and, probably, Error
   existentials in general)
-> Tuples of Errors.

The patch switches this codepath to RemoteAST, removing a bit
of code, and fixing several corner cases of existential resolution.
There are maybe other cases that are still broken, but this is
a decent enough improvement that I figure I could just go ahead
and commit.

Fixes:
<rdar://problem/41228727>
<rdar://problem/37479968>
<rdar://problem/41229005>

Adding several testcases, while here, to make sure this doesn't
regress in future.