Drop support for non-standard Swift resource directories.
Attempting to import anything from  a Swift resource directory that is
not the one used by LLDB's  embedded Swift compiler is doomed to fail;
especially so if the binary being debugged was compiled on  a different
machine.

Note that the resource dir does not encode the target platform, so
this should not affect cross-debugging.

Part of rdar://problem/45282507

(cherry picked from commit fa1c9f6ce416e14fe6117b6788bc0eb9c62dd050)