[lldb] Fix race condition in framework installation
We need the install-liblldb-stripped target to depend on the
lldb-framework target in order for the installation to be guaranteed to
behave correctly, otherwise it's possible for the lldb-framework and
install-liblldb-stripped targets to run in parallel, resulting in
temporary or partially processed files being copied into the framework.
install-liblldb already depends on lldb-framework for this reason.

Differential Revision: https://reviews.llvm.org/D53917

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@345711 91177308-0d34-0410-b5e6-96231b3b80d8