Remove Go debugger plugin
In January Davide sent an e-mail to the mailing list to suggest removing
unmaintained language plugins such as Go and Java. The plan was to have
some cool down period to allow users to speak up, however after that the
plugins were never actually removed.

This patch removes the Go debugger plugin.

The plugin can be added again in the future if it is mature enough both
in terms of testing and maintenance commitment.

Discussion on the mailing list:
http://lists.llvm.org/pipermail/lldb-dev/2018-January/013171.html

Differential revision: https://reviews.llvm.org/D54057

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@346157 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 2068e61948f32a410c40d5af3599588bbb277b27)