Remove OCaml debugger plugin
In January Davide sent an e-mail to the mailing list to suggest removing
unmaintained language plugins such as Go and Java. The plan was to have
some cool down period to allow users to speak up, however after that the
plugins were never actually removed.

This patch removes the OCaml debugger plugin.

The plugin can be added again in the future if it is mature enough both
in terms of testing and maintenance commitment.

Discussion on the mailing list:
http://lists.llvm.org/pipermail/lldb-dev/2018-January/013171.html

Differential revision: https://reviews.llvm.org/D54060

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@346159 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit ca607131c450811fa4a5ef5edc7e8557ae16ea11)