Resubmit "Fix bug in PE/COFF plugin."
The original commit was actually 2 unrelated bug fixes, but it turns
out the second bug fix wasn't quite correct, so the entire patch was
reverted.  Resubmitting this half of the patch by itself, then will
follow up with a new patch which fixes the rest of the issue in a
more appropriate way.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@346505 91177308-0d34-0410-b5e6-96231b3b80d8