Fix a use-after-free of the ABI plugin.
This was introduced in r346775.  Previously the ABI shared_ptr
was declared as a function local static meaning it would live
forever.  After the change, someone has to create a strong
reference to it or it will go away.  In this code, we were
calling ABI::FindPlugin(...).get(), so it was being immediately
destroyed and we were holding onto a dangling pointer.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@346932 91177308-0d34-0410-b5e6-96231b3b80d8