[driver] Fix --core/-c and add test
Because the optarg variable was shadowed we didn't notice we weren't
extracting the value from the option. This patch fixes that and renames
the variable to prevent this from happening in the future.

I also added two tests to check the error output for --core and --file
when the given value doesn't exist.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@347821 91177308-0d34-0410-b5e6-96231b3b80d8