Remove getopt includes from the driver
They are not needed now that we use LLVMOption for command-line parsing
thank you, Jonas).  This also allows us to avoid linking of lldbHost
into the driver which was breaking liblldb encapsulation.

(Technically, there is still a lldb/Host/windows/windows.h include which
is needed on windows, but this is a header-only wrapper for <windows.h>,
so it is not necessary to link lldbHost for that. But ideally, that
should go away too.)

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@347846 91177308-0d34-0410-b5e6-96231b3b80d8