[Expr] Fix `TestExprOptions` after r348240 on MacOS X
Summary:
r348240 assumes that an expression contains the Objective C option if
Objective C Runtime is found. But on MacOS X it seems that the test application
process always contains Objective C Runtime, so the test fails when it assumes
that the language is C++ only. Skip this part on Darwin.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@348250 91177308-0d34-0410-b5e6-96231b3b80d8