ELF: Clean up section type computation
Move code into a separate function, and replace the if-else chain with
llvm::StringSwitch.

A slight behavioral change is that now I use the section flags
(SHF_TLS) instead of the section name to set the thread-specific
property. There is no explanation in the original commit introducing
this (r153537) as to why that was done this way, but the new behavior
should be more correct.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@348936 91177308-0d34-0410-b5e6-96231b3b80d8