Support unknown unowned references to Objective-C classes.
On Darwin the Swift runtime stores unowned references to Objective-C
objects as a pointer to a struct that has the actual object pointer at
offset zero. The least significant bit of the reference pointer
indicates whether the reference refers to an Objective-C or Swift
object.

This is a property of the Swift runtime(!). In the future it
may be necessary to check for the version of the Swift runtime
(or indirectly by looking at the version of the remote
operating system) to determine how to interpret references.

This patch teaches the SwiftLanguageRuntime about this and handles
unwoned references accordingly whengetting the data of any
ValueObjectChildren (which classes always are).

<rdar://problem/42300829>