[cmake] Make libcxx(abi) a dependency when building in-tree clang for macOS
As discussed on IRC this morning, when building an in-tree clang for
testing we have to have libcxx and libcxxabi checked out. This is a
common pitfall. Not only are the resulting failures non-obvious, they
only manifest when running the test suite, *after* building lldb and
clang. By making them a hard dependency (on macOS) we fail earlier with
a more useful error message.

Differential revision: https://reviews.llvm.org/D55837

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@349539 91177308-0d34-0410-b5e6-96231b3b80d8