[lit] Make TestConvenienceVariables a cpp file
The build.py script always runs the compiler in C++ mode, regardless of
the file extension. This results in mangled names presented to the
linker which in turn cannot find the printf symbol.

While we figure out how to solve this issue I've turned the source file
into a cpp file and added extern c. This should unbreak the bots.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@349642 91177308-0d34-0410-b5e6-96231b3b80d8