Show the memory region name if there is one in the output of the "memory region" command
Prior to this change we would show the name of the section that a memory region belonged to but not its actual region name. Now we show this,. Added a test that reuses the regions-linux-map.dmp minidump file to test this and verify the correct region names for various memory regions.

Differential Revision: https://reviews.llvm.org/D55854



git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@349658 91177308-0d34-0410-b5e6-96231b3b80d8