[lldbsuite] Un-xfail tests on Windows that are now passing
This is a set of tests that were all marked as failing becuse of pr24489. The bug is not fixed (as in more of the tests that were marked this way are failing), but this set is passing. It is possible that some of them are false positives, but there's a large number of unexpectedly passing tests on Windows, so I am doing a bulk un-xfail to get the buildbot to green.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@349665 91177308-0d34-0410-b5e6-96231b3b80d8