[NativePDB] Implement ParseDeclsForContext.
This is a first step towards getting lldb-test symbols working
with the native plugin.  There is a remaining issue, which is
that the plugin expects that ParseDeclsForContext will also
create lldb symbols rather than just the decls, but the native
pdb plugin doesn't currently do this.  This will be addressed
in a followup patch.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@350243 91177308-0d34-0410-b5e6-96231b3b80d8