Rearrange bitfield to allow for more space in file_idx.
This is an alternate patch for the bug reported in
https://bugs.llvm.org/show_bug.cgi?id=39816 ("lldb doesn't show a file
of line entry for big project"). This limits the number of lines in a
source file to 32M, which I think is reasonable even for preprocessed
source inputs.

Differential Revision: https://reviews.llvm.org/D56218

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@350274 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit f55f22d191ea369618a7f690484a88e19c2e829c)