[CMake] Remove use of CLANG_MAIN_INCLUDE_DIR and SWIFT_MAIN_INCLUDE_DIR
These are only used in one place, so just inline the value. Although
CLANG_MAIN_SRC_DIR and SWIFT_MAIN_SRC_DIR are gated by having access to the
source trees, and so CLANG_MAIN_INCLUDE_DIR and SWIFT_MAIN_INCLUDE_DIR would just
be "/include". This behavior is preserved. Additionally, we almost always have
access to the source trees for Clang and Swift, so this shouldn't happen often if
at all.