[Reproducers] SBReproducer framework: Capture & Replay
This is part two of the reproducer instrumentation framework. It
contains the code to capture and replay function calls. The main user of
this framework will be the SB API layer.

For all the details refer to the RFC on the mailing list:
http://lists.llvm.org/pipermail/lldb-dev/2019-January/014530.html

Differential revision: https://reviews.llvm.org/D56322

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@353324 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 6e7ade8b386731ab5d2efb46e98e780b4c545788)