Breakpad: temporarily XFAIL line-table tests
The tests are failing on windows because the paths in the symbol file
are parsed using the host path style. I'm working on a patch to have
SymbolFileBreakpad auto-detect the correct path style (similar to dwarf
r351328).

I originally wanted to make this a part of the initial line-table patch,
but then I simply forgot.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@353410 91177308-0d34-0410-b5e6-96231b3b80d8