Set Swift ABI bit variable when creating a SwiftRuntime
The Swift metadata for classes uses a different bit to identify themselves in the pre-
and post-stable ABI eras. We need to be able to debug both, so the code in the MetadataReader
needs to be configured at runtime.

This commit needs to be paired with one in the swift sources to allow the configuration
of this value. This commit shouldn't break anything without it though.

rdar://problem/47935062