Revert "Fix one more string/bytes issue in lldb-server tests"
It looks like I was too hasty to submit the previous patch. It does fix
some tests on python3, but it also breaks one tests with python2.

This happens because the gdb-remote protocol can sometimes (but not very
often) contain binary data, and attempting to parse this as utf8
characters fails.

This reverts commit r353944.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@353945 91177308-0d34-0410-b5e6-96231b3b80d8