Fix AST generated for a class template to connect the class inside a class template back to the template.
Previously, when the ASTImporter imported the class, it didn't know that
it was the pattern of a class template, so made the class a name lookup
result for the name of the template, resulting in ambiguity errors when
naming the template.

Due to a clang bug (fixed in r354091, reverted and soon to be
re-committed), ambiguity errors between a template and a non-template
were previously not diagnosed. Once r354091 is re-committed, this will
be covered by existing lldb tests.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@354173 91177308-0d34-0410-b5e6-96231b3b80d8