Remove unnecessary demangling operation (NFC)
This extra call to the demangler doesn't affect the performance of C++
because the result is being cached anyway; but I'm working on a patch
to the Swift branch that uses extra contextual information to provide
a more accurate demangling result. In that case this call would be
extra and unnecessary work.

Differential Revision: https://reviews.llvm.org/D58720

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@355042 91177308-0d34-0410-b5e6-96231b3b80d8