[Reproducers] Implement log_append for function pointers.
Changing the type in the DUMMY macro to void* doesn't actually fix the
build error, because the argument type is deducted from the template (as
opposed to when serializing through the instrumentation framework, where
this would matter). Instead I've added a proper instance of log_append
that takes function pointers and logs their address.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@355863 91177308-0d34-0410-b5e6-96231b3b80d8