[Reproducers] Support capturing a reproducer without an explicit path.
Tablegen doesn't support options that are both flags and take values as
an argument. I noticed this when doing the tablegen rewrite, but forgot
that that affected the reproducer --capture flag.

This patch makes --capture a flag and adds --capture-path to specify a
path for the reproducer. In reality I expect this to be mostly used for
testing, but it could be useful nonetheless.

Differential revision: https://reviews.llvm.org/D59238

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@355936 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 8c2e36ab579e6574e497513cec82aac1008cb6b1)