[Reproducers] Stop recording instead of deallocating
The command interpreter holds a pointer to a DataRecorder. After
generating the reproducer, we deallocated all the DataRecorders, causing
the command interpreter to hold a non-null reference to an invalid
object.

This patch changes the behavior of the command provider to stop the
DataRecorders when a reproducer is generated, rather than deallocating
them.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@355940 91177308-0d34-0410-b5e6-96231b3b80d8