Correcting some comments in PdbIndex.cpp [NFC]
ICF can cause multiple symbols to start at the same virtual address.
I plan to handle this shortly, but I wanted to correct the comment for
now.

Deleted an obsolete comment about adjusting the offset for the magic
number at the beginning of the debug info stream.  This adjustment is
handled at a lower level now.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@355943 91177308-0d34-0410-b5e6-96231b3b80d8