[Python] Define empty SWIG wrapper for unit testin"
The python plugin uses wrappers generated by swig. For the symbols to be
available, we'd need to link against liblldb, which is not an option
because the symbols could conflict with the static library we are
testing. Instead we define the symbols ourselves in the unit test.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@356971 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit ade26efec4f9607efe44c47e2a6d6bbdfa7df6eb)