[Python] Remove readline module
Todd added this empty readline module to workaround an issue with an old
version of Python on Ubuntu in 2014 (18841). In the meantime, libedit
seems to have fixed the underlying issue, and indeed, I wasn't able to
reproduce this.

Differential revision: https://reviews.llvm.org/D59972

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@357277 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 1b2170d0116d52a219574780e7eb01043c3712e1)