[CMake] Remove Apple-specific version logic.
We were using the LLDB-Info.plist as the canonical holder of the
version number, but there is really no good reason to do this. If
anything the plist should be generated using the information provided
to CMake.

For now just remove the logic extracting the version from the plist
and rely on LLDB_VERSION_STRING.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@358604 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 0474b0eafc931d139bc4f07904fb0959297c4ab0)