[Reproducers] Fix lifetime issue
Deallocating the data recorder in during the ::Keep() operation causes
problems down the line when exiting the debugger. The command
interpreter still holds a pointer to the now deallocated object and has
no way to know it no longer exists. This is exactly what the m_record
flag was meant for, although it wasn't hooked up properly either.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@358916 91177308-0d34-0410-b5e6-96231b3b80d8